core/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.1
django-modeltranslation==0.18.11
djangorestframework==3.14.0
mysqlclient==2.2.0
Pillow==10.1.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
